modificato tipo colonna durata messaggio

+++ FomMonitoringCore/DAL_SQLite/message.cs
@@ -16,7 +16,7 @@ namespace FomMonitoringCore.DAL_SQLite
     {
         public int Id { get; set; }
         public Nullable<System.DateTime> Time { get; set; }
-        public Nullable<int> TimeSpanDuration { get; set; }
+        public Nullable<long> TimeSpanDuration { get; set; }
         public Nullable<int> Code { get; set; }
         public string Params { get; set; }
         public string Operator { get; set; }
